crash: require python 3.6
Convert the project to use Python 3.6 syntax:
- print() as function
- except <exception> as <name>
- absolute imports
- no more __future__ imports
- int() instead of long()
- no keying off of sys.version.major

This also enables the use of type checking and f-strings but those conversions
will happen separately.

+++ contrib/lsbtrfs-cmd.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/python3
 # vim:set shiftwidth=4 softtabstop=4 expandtab textwidth=79:
 
 path = "/lib/modules/4.4.20-6.gd2e08c5-default/kernel/fs/btrfs/btrfs.ko"
@@ -37,7 +37,7 @@ class LSBtrfs(CrashCommand):
             if sb['s_type']['name'].string() == "btrfs":
                 fs_info = gdb.Value(sb['s_fs_info']).cast(btrfs_fs_info_type.pointer())
 
-                u = long(0)
+                u = 0
                 for i in range(0, 16):
                     u <<= 8
                     u += int(fs_info['fsid'][i])
